Suscripciones arreglada

+++ TP-PAV-3K02/Modelos/Sucripciones.cs
@@ -10,7 +10,7 @@ namespace TP_PAV_3K02.Modelos
 {
     public class Suscripcion
     {
-        public int nro_doc { get; set; }
+        public long nro_doc { get; set; }
         public int cod_TipoDoc { get; set; }
         public int cod_int { get; set; }
         public DateTime fecha_inicio { get; set; }
